List plug-in directories when plug-in call fails
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS/trunk@7011 212acab6-be3b-0410-9dea-997c60f758d6

+++ Source/exehead/fileform.h
@@ -516,7 +516,7 @@ typedef struct {
 
 #define REGROOTVIEW32 0x40000000
 #define REGROOTVIEW64 0x20000000
-#define REGROOTVIEWTOSAMVIEW(rv) ( ((UINT_PTR)(rv)&(REGROOTVIEW32|REGROOTVIEW64)) >> 21 ) // REGROOTVIEWxx to KEY_WOW64_xxKEY
+#define REGROOTVIEWTOSAMVIEW(rv) ( (UINT)(((UINT_PTR)(rv)&(REGROOTVIEW32|REGROOTVIEW64)) >> 21) ) // REGROOTVIEWxx to KEY_WOW64_xxKEY
 #define IsRegRootkeyForcedView(hKey) ( ((UINT_PTR) (hKey) & (REGROOTVIEW32|REGROOTVIEW64)) )
 #define MAKEREGROOTVIEW(r, fv) ( (HKEY) ((UINT_PTR)(r) | (fv)) )
 #define HKSHCTX ( (HKEY) 0 ) // Converted to HKCU or HKLM by GetRegRootKey
